Refactored TwoTouchChangeSizeAndMoveHandler and SaveTextBox/EditTextBox Handlers into their own classes.
